+++ EnhanceQoLMythicPlus/Locales/zhTW.lua
@@ -76,6 +76,7 @@ L["Teleports"] = "傳送"
 L["teleportEnabled"] = "啟用傳送框架"
 L["DungeonCompendium"] = "地城傳送指南"
 L["teleportsEnableCompendium"] = "啟用地城傳送指南"
+L["teleportCompendiumHeadline"] = "隱藏特定資料片的傳送"
 
 L["teleportsHeadline"] = "在PvE視窗中添加一個包含地城傳送的框架"
 L["portalHideMissing"] = "隱藏缺失的傳送"